pyproject.toml
[project]
name = "wednesday-solutions-assignment"
version = "0.1.0"
description = "Lead Generation Automation System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.2",
    "tenacity>=8.2.3", # For retry mechanisms
    "beautifulsoup4>=4.12.2", # For web scraping
    "openai>=1.3.7", # For AI message generation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "isort>=5.12.0",
    "pytest-cov>=4.1.0",
]
